Monotone-Parent: 58bc704a5111b6f454c19eeabc083831a734ce1e
Monotone-Parent: 9c06cc61730dff177d1315d9321971000ee5b90b
Monotone-Revision: 4c6f15da587adb49bf5eec3b0a79129284da706b

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2009-12-21T17:46:29
Monotone-Branch: org.NetBSD.atf.src

+++ tests/atf/atf-c++/t_fs.cpp
@@ -600,7 +600,7 @@ ATF_TEST_CASE_BODY(temp_file_fd)
 
         path tmpl("tempfile.XXXXXX");
         temp_file tf(tmpl);
-        ATF_CHECK_EQUAL(msg.length(),
+        ATF_CHECK_EQUAL(static_cast< int >(msg.length()),
                         ::write(tf.fd(), msg.c_str(), msg.length()));
         tf.close();
